Branch off a separate test for method renames
